+++ skyfield/__init__.py
@@ -5,5 +5,5 @@ functions and classes there.  See the ``documentation`` directory inside
 the source code, as well as the http://rhodesmill.org/skyfield/ site!
 
 """
-VERSION = (1, 34)
+VERSION = (1, 35)
 __version__ = '.'.join(map(str, VERSION))